slight refactor, reorg of files
